requirements.txt
# PDF processing
PyMuPDF==1.24.4
pypdf>=3.0.0  # For CaseLawPipeline

# Text processing
regex
unicodedata2
python-dateutil

# OCR (optional)
pytesseract
Pillow

# General utilities
tqdm
python-dotenv

# Graph database
neo4j>=5.14.0

# LangChain ecosystem
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Embeddings - sentence-transformers
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.35.0

# NER - BERT-based and spaCy fallback
# transformers already included above for BERT NER
spacy>=3.7.0

# Vector search (for testing/fallback)
faiss-cpu>=1.7.4

# Additional ML utilities
numpy>=1.24.0
scikit-learn>=1.3.0

# Case Law Pipeline
google-genai>=0.2.0  # Gemini 1.5 Flash for semantic segmentation (replaces deprecated google-generativeai)
